Refactor styled components
