+++ components/styled/weather-detail.js
@@ -1,0 +1,47 @@
+import styled from 'styled-components'
+
+export const HourlyWeatherItem = styled.div`
+    display: inline-block;
+    text-align: center;
+    width: 16%;
+`
+export const HourlyDate = styled.span`
+    font-size: 0.85em;
+`
+export const HourlyTemp = styled.span`
+    font-size: 1.1em;
+    font-weight: bold;
+`
+export const WeatherDetail = styled.div`
+    width: 100%;
+    position: relative;
+`
+export const WeatherDetailHeader = styled.h1`
+    font-weight: bold;
+    line-height: 1.5;
+`
+export const WeatherDetailDate = styled.span`
+    font-weight: bold;
+    font-size: 1.1em;
+    line-height: 1.5;
+`
+export const WeatherDetailCaption = styled.div`
+    font-size: 0.85em;
+    color: #717171;
+`
+export const WeatherDetailTemp = styled.div`
+    font-size: 5em;
+    font-weight: bold;
+    text-align: center;
+`
+export const WeatherDetailCondition = styled.div`
+    font-size: 1.25em;
+    text-align: center;
+    padding-top: 8px;
+`
+export const WeatherSubHeader = styled.h5`
+    font-size: 0.85em;
+    color: #717171;
+    margin-bottom: 20px;  
+    text-transform: uppercase;  
+`
